+++ server/src/test/java/org/opensearch/indices/replication/SegmentReplicationTargetServiceTests.java
@@ -86,6 +86,7 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
             checkpoint,
             indexShard,
             replicationSource,
+            false,
             new SegmentReplicationTargetService.SegmentReplicationListener() {
                 @Override
                 public void onReplicationDone(SegmentReplicationState state) {
@@ -115,6 +116,7 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
             checkpoint,
             indexShard,
             replicationSource,
+            false,
             new SegmentReplicationTargetService.SegmentReplicationListener() {
                 @Override
                 public void onReplicationDone(SegmentReplicationState state) {
@@ -141,7 +143,7 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
     public void testAlreadyOnNewCheckpoint() {
         SegmentReplicationTargetService spy = spy(sut);
         spy.onNewCheckpoint(indexShard.getLatestReplicationCheckpoint(), indexShard);
-        verify(spy, times(0)).startReplication(any(), any(), any());
+        verify(spy, times(0)).startReplication(any(), any(), any(), any());
     }
 
     public void testShardAlreadyReplicating() throws InterruptedException {
@@ -151,6 +153,7 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
             checkpoint,
             indexShard,
             replicationSource,
+            false,
             mock(SegmentReplicationTargetService.SegmentReplicationListener.class)
         );
         // Create a Mockito spy of target to stub response of few method calls.
@@ -173,20 +176,20 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
 
         // wait for the new checkpoint to arrive, before the listener completes.
         latch.await(30, TimeUnit.SECONDS);
-        verify(serviceSpy, times(0)).startReplication(eq(aheadCheckpoint), eq(indexShard), any());
+        verify(serviceSpy, times(0)).startReplication(eq(aheadCheckpoint), eq(indexShard), any(), any());
     }
 
     public void testNewCheckpointBehindCurrentCheckpoint() {
         SegmentReplicationTargetService spy = spy(sut);
         spy.onNewCheckpoint(checkpoint, indexShard);
-        verify(spy, times(0)).startReplication(any(), any(), any());
+        verify(spy, times(0)).startReplication(any(), any(), any(), any());
     }
 
     public void testShardNotStarted() throws IOException {
         SegmentReplicationTargetService spy = spy(sut);
         IndexShard shard = newShard(false);
         spy.onNewCheckpoint(checkpoint, shard);
-        verify(spy, times(0)).startReplication(any(), any(), any());
+        verify(spy, times(0)).startReplication(any(), any(), any(), any());
         closeShards(shard);
     }
 
@@ -197,9 +200,9 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
         ArgumentCaptor<SegmentReplicationTargetService.SegmentReplicationListener> captor = ArgumentCaptor.forClass(
             SegmentReplicationTargetService.SegmentReplicationListener.class
         );
-        doNothing().when(spy).startReplication(any(), any(), any());
+        doNothing().when(spy).startReplication(any(), any(), any(), any());
         spy.onNewCheckpoint(aheadCheckpoint, spyShard);
-        verify(spy, times(1)).startReplication(any(), any(), captor.capture());
+        verify(spy, times(1)).startReplication(any(), any(), any(), captor.capture());
         SegmentReplicationTargetService.SegmentReplicationListener listener = captor.getValue();
         listener.onFailure(new SegmentReplicationState(new ReplicationLuceneIndex()), new OpenSearchException("testing"), true);
         verify(spyShard).failShard(any(), any());
@@ -215,11 +218,11 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
         // Starting a new shard in PrimaryMode.
         IndexShard primaryShard = newStartedShard(true);
         IndexShard spyShard = spy(primaryShard);
-        doNothing().when(spy).startReplication(any(), any(), any());
+        doNothing().when(spy).startReplication(any(), any(), any(), any());
         spy.onNewCheckpoint(aheadCheckpoint, spyShard);
 
         // Verify that checkpoint is not processed as shard is in PrimaryMode.
-        verify(spy, times(0)).startReplication(any(), any(), any());
+        verify(spy, times(0)).startReplication(any(), any(), any(), any());
         closeShards(primaryShard);
     }
 
@@ -244,10 +247,10 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
         ArgumentCaptor<SegmentReplicationTargetService.SegmentReplicationListener> captor = ArgumentCaptor.forClass(
             SegmentReplicationTargetService.SegmentReplicationListener.class
         );
-        doNothing().when(spy).startReplication(any(), any(), any());
+        doNothing().when(spy).startReplication(any(), any(), any(), any());
         spy.onNewCheckpoint(newCheckpoint, spyShard);
         spy.onNewCheckpoint(anotherNewCheckpoint, spyShard);
-        verify(spy, times(1)).startReplication(eq(newCheckpoint), any(), captor.capture());
+        verify(spy, times(1)).startReplication(eq(newCheckpoint), any(), any(), captor.capture());
         verify(spy, times(1)).onNewCheckpoint(eq(anotherNewCheckpoint), any());
         SegmentReplicationTargetService.SegmentReplicationListener listener = captor.getValue();
         listener.onDone(new SegmentReplicationState(new ReplicationLuceneIndex()));
@@ -262,6 +265,7 @@ public class SegmentReplicationTargetServiceTests extends IndexShardTestCase {
             checkpoint,
             indexShard,
             replicationSource,
+            false,
             mock(SegmentReplicationTargetService.SegmentReplicationListener.class)
         );
         final SegmentReplicationTarget spy = Mockito.spy(target);